Move relevant-filter into import-export folder
