+++ src/ggrc-client/js/components/import-export/export.js
@@ -3,7 +3,7 @@
   Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 */
 
-import '../relevant_filters';
+import './relevant-filter';
 import './export-group';
 import csvExportTemplate from './templates/csv-export.mustache';
 import {confirm} from '../../plugins/utils/modals';
